multi_automation_app/requirements.txt
streamlit
pywhatkit
twilio
pillow
requests
beautifulsoup4
googlesearch-python
